requirements.txt
streamlit
pandas
yfinance
matplotlib.pyplot
datetime
prophet
